renamed models to be more logical and deleted competition coach list view
