+++ app/baseball/models/__init__,.py
@@ -1,6 +1,6 @@
 from .coach import Coach
-from .coach_competition_stats import CoachCompetitionStats
+from .competition_coach import CompetitionCoach
 from .player import Player
-from .player_competition_stats import PlayerCompetitionStats
+from .competition_player import CompetitionPlayer
 from .competition import Competition
 from .organization import Organization
