+++ app/baseball/serializers/coach_serializer.py
@@ -1,12 +1,12 @@
 from rest_framework import serializers
 
 from ..models.coach import Coach
-from ..models.coach_competition_stats import CoachCompetitionStats
+from ..models.competition_coach import CompetitionCoach
 
 class StatsByCompetitionField (serializers.RelatedField):
     """Custom relational field for a coach's stats by competition.
     """
-    def to_representation(self, value: CoachCompetitionStats):
+    def to_representation(self, value: CompetitionCoach):
         """Overwritten method that shows how each CoachCompetitionStats in the stats_by_competition
         will be displayed."""
         return f'{value.competition.name}'
