+++ app/baseball/serializers/competition_serializer.py
@@ -1,20 +1,20 @@
 from rest_framework import serializers
 
 from ..models.competition import Competition
-from ..models.coach_competition_stats import CoachCompetitionStats
-from ..models.player_competition_stats import PlayerCompetitionStats
+from ..models.competition_coach import CompetitionCoach
+from ..models.competition_player import CompetitionPlayer
 from ..models.competition_team import CompetitionTeam
     
 class CompetitionCoachField (serializers.RelatedField):
     """Custom relational field for a competition's coaches.
     """
-    def to_representation(self, value: CoachCompetitionStats):
+    def to_representation(self, value: CompetitionCoach):
         return f'{value.coach.first_name} {value.coach.last_name}'
 
 class CompetitionPlayerField (serializers.RelatedField):
     """Custom relational field for a competition's players.
     """
-    def to_representation(self, value: PlayerCompetitionStats):
+    def to_representation(self, value: CompetitionPlayer):
         return f'{value.player.first_name} {value.player.last_name}'
 
 class CompetitionTeamsField (serializers.RelatedField):
