+++ app/baseball/serializers/player_serializer.py
@@ -1,12 +1,12 @@
 from rest_framework import serializers
 
 from ..models.player import Player
-from ..models.player_competition_stats import PlayerCompetitionStats
+from ..models.competition_player import CompetitionPlayer
 
 class StatsByCompetitionField (serializers.RelatedField):
     """Custom relational serializer for a player's stats by competition.
     """
-    def to_representation(self, value: PlayerCompetitionStats):
+    def to_representation(self, value: CompetitionPlayer):
         """Overwritten method that shows how each PlayerCompetitionStats in the stats_by_competition
         will be displayed."""
         return f'{value.competition.name}'
